役割ベースのログ処理構成へ整理する
