--- src/app.rs
@@ -1,2 +1,0 @@
-pub mod commands;
-pub mod run;
