--- src/app/commands.rs
@@ -1,3 +1,0 @@
-pub mod cat;
-pub mod shared;
-pub mod watch;
