--- src/core.rs
@@ -1,4 +1,0 @@
-pub mod filter;
-pub mod highlight;
-pub mod matcher;
-pub mod processor;
